Moved dosimetry files to dosimetry folders
