+++ cmake/unit_testing.cmake
@@ -25,11 +25,9 @@ add_subdirectory( src/ACEtk/block/ReactionNumberBlock/test )
 add_subdirectory( src/ACEtk/block/ReactionQValueBlock/test )
 add_subdirectory( src/ACEtk/block/FrameAndMultiplicityBlock/test )
 add_subdirectory( src/ACEtk/block/CrossSectionData/test )
-add_subdirectory( src/ACEtk/block/DosimetryCrossSectionData/test )
 add_subdirectory( src/ACEtk/block/CrossSectionBlock/test )
 add_subdirectory( src/ACEtk/block/TabulatedFissionMultiplicity/test )
 add_subdirectory( src/ACEtk/block/FissionMultiplicityBlock/test )
-add_subdirectory( src/ACEtk/block/DosimetryCrossSectionBlock/test )
 add_subdirectory( src/ACEtk/block/EquiprobableAngularBins/test )
 add_subdirectory( src/ACEtk/block/EquiprobableOutgoingEnergyBins/test )
 add_subdirectory( src/ACEtk/block/TabulatedAngularDistribution/test )
@@ -89,6 +87,9 @@ add_subdirectory( src/ACEtk/photoatomic/HeatingNumbersBlock/test )
 add_subdirectory( src/ACEtk/photoatomic/ComptonProfile/test )
 add_subdirectory( src/ACEtk/photoatomic/ComptonProfileBlock/test )
 add_subdirectory( src/ACEtk/photoatomic/PhotoelectricCrossSectionBlock/test )
+
+add_subdirectory( src/ACEtk/dosimetry/CrossSectionData/test )
+add_subdirectory( src/ACEtk/dosimetry/CrossSectionBlock/test )
 
 add_subdirectory( src/ACEtk/electron/ElectronShellBlock/test )
 add_subdirectory( src/ACEtk/electron/ElectronSubshellBlock/test )
